requirements.txt
bottle
Mako
MarkupSafe
psycopg2-binary
redis
pyjwt
